+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Features
 ^^^^^^^^
 - Pushed Tornado to 5.x and unpinned Flask
 - Added a favicon
+- new primitive `s_bytes` which fuzzes an arbitrary length binary value (similiar to `s_string`)
 - We are now using `Black` for code style standardization
 
 Fixes